pyproject.toml
[project]
name = "table2json"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Chansong You", email = "dbcksthd13@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pycryptodome>=3.21.0",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
    "tree-sitter>=0.24.0",
    "tree-sitter-c-sharp>=0.23.1",
]

[project.scripts]
table2json = "table2json.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
